migrate all arenas to seperate files
