+++ src/RobotSimulator/Areanas/bowling.ts
@@ -1,0 +1,107 @@
+import { ArenaConfig, ChallengeConfig } from "./base";
+
+export const arenas = [arena];
+export const challenges = [challenge];
+
+function arena(): ArenaConfig {
+  const arenaConfig: ArenaConfig = {
+    name: "Bowling",
+    worldConfig: {
+      zLength: 60,
+      xLength: 20,
+      walls: [],
+      camera: {
+        position: {
+          x: 0,
+          y: 10,
+          z: 10,
+        },
+      },
+    },
+    ballSpecs: [{ type: "ball", radius: 1, initialPosition: { x: 0, y: -5 } }],
+    coneSpecs: [
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: 1.5, y: -25 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: -1.5, y: -25 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: 4.5, y: -25 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: -4.5, y: -25 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: 0, y: -22 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: 3, y: -22 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: -3, y: -22 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: 1.5, y: -19 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: -1.5, y: -19 },
+      },
+      {
+        type: "cone",
+        baseColor: 0x980000,
+        height: 5,
+        radius: 0.5,
+        initialPosition: { x: 0, y: -16 },
+      },
+    ],
+  };
+
+  return arenaConfig;
+}
+
+function challenge(): ChallengeConfig {
+  const challengeConfig: ChallengeConfig = {
+    name: "Bowling Challenge",
+    startPosition: { x: 0, y: 10 },
+    arenaConfig: arena(),
+  };
+
+  return challengeConfig;
+}
